improves compatibility with VS 2019 and lower compute capabilities and fixes string include

+++ src/GksGpu/Definitions/AccumulatorDataType.h
@@ -40,7 +40,7 @@
 //! precision accumulator can prevent undeterministic behavior
 //! that stems from the order of addition to the accumulator.
 
-//typedef float realAccumulator;
-typedef double realAccumulator;
+typedef float realAccumulator;
+//typedef double realAccumulator;
 
 #endif
